pyproject.toml
[tool.poetry]
name = "yandex-logs-api"
version = "0.1.0"
description = "Yandex Logs Api client"
authors = ["Vladimir Trifonov <3fonov@gmail.com>"]
readme = "README.md"
packages = [{include = "yandex_logs_api"}]

[tool.poetry.dependencies]
python = "^3.10"
tenacity = "^8.2.2"
aiohttp = "^3.8.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
autopep8 = "^2.0.2"
pytest-asyncio = "^0.20.3"
pytest-env = "^0.8.1"
mypy = "^1.1.1"
flake8 = "^6.0.0"
mimesis = "^7.0.0"
python-dotenv = "^0.21.1"
vcrpy = "^4.2.1"
pytest-recording = "^0.12.2"
flake8-global-variables = {git = "https://github.com/i02sopop/flake8-global-variables"}
flake8-grug = {git = "https://github.com/c0ntribut0r/flake8-grug"}
black = "^23.3.0"
flake8-json = "^21.7.0"
flake8-adjustable-complexity = "^0.0.6"
flake8-annotations-coverage = "^0.0.6"
flake8-annotations = "^3.0.0"
flake8-builtins = "^2.1.0"
flake8-cognitive-complexity = "^0.1.0"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.11.1"
flake8-expression-complexity = "^0.0.11"
flake8-functions = "^0.0.7"
flake8-simplify = "^0.20.0"
flake8-scream = "^0.1.0"
pep8-naming = "^0.13.3"
flake8-pylint = "^0.2.0"
flake8-bugbear = "^23.3.23"
flake8-pie = "^0.16.0"
flake8-print = "^5.0.0"
flake8-use-pathlib = "^0.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
